python/nrf_sniffer/requirements.txt
pyserial >= 3.5
psutil
